registry/auth: add support for client auth handling
Currently, for registry operations requiring auth (such as pull/push),
the client is expected to preemptively send credentials it believes will
be appropriate/required alonside with the image pull/push request to the
daemon, which the daemon will then use to negotiate any challenges (in
the form of `WWW-Authenticate` headers) it encounters while
communicating with the registry.

This means that, as it stands, the daemon has no avenue of communication
to request credentials as-needed, during a registry operation – all the
credentials must be sent up-front by the client (who without foresight
must somehow know credentials will actually be necessary – which poses
challenges when paired with mirrors for private registries, etc. since
in cases like these the client won't necessarily know where the engine
will actually be pulling images from, and we also don't want to just
send client credentials to any host).

This patch introduces a new way of handling auth during push/pull –
following what has been discussed in
https://github.com/moby/moby/issues/38591, instead of negotiating auth
on behalf of the client (using the provided credentials), the daemon
will instead send a `401` back to the client alonside with the received
`WWW-Authenticate` header when a challenge is received. The client can
then handle the challenge however it sees fit, pick the credentials it
wants to use (which addresses the issue with private registry mirrors)
and then initiate another pull with the relevant credentials.

This new auth handling flow is also better suited for situations where
oauth is being used to authenticate against the registry – in case it's
necessary to do a refresh flow to fetch a new access token, the client
can handle the refresh flow, store the new refresh+access tokens, and
start the pull with the new fresh credentials.

Signed-off-by: Laura Brehm <laurabrehm@hey.com>
Signed-off-by: Sebastiaan van Stijn <github@gone.nl>