api/ping: include engine features in response
Add an `Engine-Features` header to the `/_ping` response to allow
clients to know what features they can expect the engine they're
negotiating with support.

Currently this is only used to advertise whether the containerd image
store integration is enabled or not – this is useful for situations like
the CLI's `docker images --tree` (although in this case the output is
ultimately mostly fine even if it doesn't include any manifests) or for
things such as the client auth handling, which would only be implemented
for pushing/pulling with the containerd store – without this, a client
would have to attempt to use the new features, get a "not implemented"
error back, and retry.

Signed-off-by: Laura Brehm <laurabrehm@hey.com>