api/ping: add support for querying engine capabilities
With moving towards using the containerd image store (as opposed to the
"legacy" graphdrivers), some new daemon features will only be
implemented for the containerd image store.

A client wanting to make use of new features would find it useful to be
able to know what features are available without having to attempt to
use them, getting an error back, and having to do another roundtrip.

This commit adds a new query parameter to the (`GET`) Ping endpoint
`/_ping?capabilities=1` that, if set, prompts the daemon to reply with a
JSON representation of it's capabilities

Since clients will generally already by hitting the `/_ping` endpoint in
order to negotiate API versions, this provides a way for clients to know
which features are supported without extra roundtrips.

This commit also adds a capability `registry-client-auth`, that
represents whether the engine supports the new client auth handling
protocol (which will only be supported, for now, if the containerd image
store integration is enabled) – without this, a client would have to
attempt to use the new features, get a "not implemented" error back, and
retry.

Signed-off-by: Laura Brehm <laurabrehm@hey.com>