api/ping: add support for querying engine features
With moving towards using the containerd image store (as opposed to the
"legacy" graphdrivers), some new daemon features will only be
implemented for the containerd image store.

A client wanting to make use of new features would find it useful to be
able to know what features are available without having to attempt to
use them, getting an error back, and having to do another roundtrip.

This commit adds a new query parameter to the (`GET`) Ping endpoint
`/_ping?features=v1` that, if set, prompts the daemon to reply with a
JSON representation of it's configured features.

Since clients will generally already by hitting the `/_ping` endpoint in
order to negotiate API versions, this provides a way for clients to know
which features are supported without extra roundtrips.

Currently this is only used to advertise whether the containerd image
store integration is enabled or not – this is useful for situations like
the CLI's `docker images --tree` (although in this case the output is
ultimately mostly fine even if it doesn't include any manifests) or for
things such as the client auth handling, which would only be implemented
for pushing/pulling with the containerd store – without this, a client
would have to attempt to use the new features, get a "not implemented"
error back, and retry.

Signed-off-by: Laura Brehm <laurabrehm@hey.com>