daemon/config: add default features to config
In 1cda161e6abd8157a4b37acc250bddc418df1227, we introduced support for a
new query parameter in the Ping endpoint (`/_ping?features=v1`) which
allows clients to query the engine for configured features, such as the
containerd image store integration.

However, features such as the containerd image store integration will
eventually by enabled by default, and so some newer daemons might have
the containerd image store integration enabled without explicitly
configuring it in their config.

This poses an issue for older clients communicating with newer daemons,
since they might assume the daemon does not support some feature when
it's actually just on by default and as such not explicitly enabled.

We'd normally address an issue like this by conditionally adding older
features that are now on-by-default to responses based on the API
version the client is using, but since the `/_ping` endpoint isn't
versioned, this is not a viable solution.

This commit adds a `DefaultFeatures` var to `daemon/config` that keeps
track of supported daemon features and their default value, and uses
this to update the daemon's config field `Features` to explicitly add
features + their default value if they are not explicitly configured,
so that we can keep including these in the features list when queried
even after they are on-by-default.

Signed-off-by: Laura Brehm <laurabrehm@hey.com>