Dockerfile: make sure modprobe is available in dev-container
Depending on the host configuration, the `br_netfilter` module may not
be loaded by default. In this situation, docker will try to load the module
through `modprobe`.

Older versions of docker would silently ignore failing to do so, log a message,
and continue;

    time="2024-11-29T20:04:58.538404376Z" level=warning msg="Running modprobe bridge br_netfilter failed with message: , error: exec: \"modprobe\": executable file not found in $PATH"

However, starting with db25b0dcd0461802289e962aa0df3abd323d1994 and
264c15bfc427d1321c5b302e48e16d113b06ef92, we now produce an error, which
in our CI situation caused tests to fail:

    === FAIL: libnetwork/drivers/bridge TestCreateFullOptions (0.04s)
    time="2024-11-29T19:03:44Z" level=error msg="Running modprobe br_netfilter failed with message: " error="exec: \"modprobe\": executable file not found in $PATH"
        bridge_linux_test.go:280: Failed to create bridge: loadBridgeNetFilterModule failed: cannot restrict inter-container communication: modprobe br_netfilter failed: exec: "modprobe": executable file not found in $PATH

This patch updates the Dockerfile to make sure `modprobe` is available in
the dev-container, which is the environment in which we run out tests.

Following this, we should;

- Check our packaging to make sure `kmod` / `modprobe` is listed as a dependency
- Look for alternatives / fallbacks for situations where `modprobe` is not
  available.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>