registry: deprecate APIEndpoint.TrimHostName
This field was added in 19515a7ad859b28c474d81e756ac245afcd968e3, but looks
to be always set for endpoints used, so we can trim remote names unconditionally.

This option was added for possible future expansion, allowing registry-
mirrors to get the full reference of the image (including domain-name),
for them to host a mirror for multiple upstreams on the same registry.

That approach will unlikely be implemented, and containerd has a different
approach for this, where the reference to the original registry is passed
through a query parameter instead.

The field is unlikely used outside of our codebased, but deprecating it
before removal just in case.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>