registry: Service.lookupV2Endpoints: add arg to skip mirrors
This function unconditionally constructed endpoints for mirrors when
requesting endpoints for the default (Docker Hub) registry. Doing so
involves validating the config, which involves;

- parsing the hostname
- constructing TLS config
- performing a DNS lookup to resolve the host's IP address and matching
  it against CIDR masks for insecure registries.

When looking up push endpoints or endpoints to consider for authentication,
mirror endpoints were discarded to prevent sending credentials of the upstream
registry to a mirror.

This patch adds a "includeMirrors" argument to skip constructing endpoints
for mirrors when not needed. While at it, also removing named output variables,
as they didn't add much.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>