daemon: remove Daemon.NetworkControllerEnabled
This function was added in eb982e7c00192c8306f9c420fb469f087c7b161d, at
which time networking was not yet implemented for Windows, resulting
in a panic when trying to call network-related endpoints.

That's no longer the case, so we should be able to add network-endpoints
unconditionally.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>