Include iface routes to unspecified addrs when looking for default gws
In L3 modes, the ipvlan driver can't set up a default gateway with
a next hop address, because there's no L2 for it to resolve the
gateway IP into a MAC address.

Instead, it sets up a route to 0.0.0.0 or [::] that's connected to
the network's interface. The end result is the same - the container
has a default route.

So, include those routes when searching for endpoints that can act
as a container's default gateway.

Signed-off-by: Rob Murray <rob.murray@docker.com>