builder: don't fall back to defaultKeepStorage when set to zero
commit b08ff8120466cb646637767cccb70ccce7a06b4b updated this code to
only produce an error if an invalid value was set by the user, and to
avoid errors on empty values.

However, the intent of this code was to allow `0` as a valid value for
cases where gc is to be handled through other properties / filters.

This patch only sets the default if no value was set by the user, but
doesn't set the default if a value is set by the user, but zero.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>