pkg/archive: nosysFileInfo: implement tar.FileInfoNames to prevent lookups
commit e9bbc41dd146d692e28660a392b068c9c112f2ad removed our fork of
pkg/archive that was in place to mitigate CVE-2019-14271. As part of that
change, a nosysFileInfo type was added to prevent tar.FileInfoHeader from
looking up user- and group-names.

A proposal was pending in go https://go.dev/issue/50102 to define an
interface for implementing custom lookup functions to be implemented,
and disable go's builtin lookup. That proposal was accepted, and is now
implemented in go1.23.

Thia patch makes the nosysFileInfo implement the tar.FileInfoNames interface
to prevent tar.FileInfoHeader from performing its own lookups. While the
mitigation implemented in e9bbc41dd146d692e28660a392b068c9c112f2ad should
already prevent this from happening, implementing the interface does not
cost us much and is complementary to the existing mitigation.

With this patch in place, we can consider removing the mitigation added
in a316b10dab79d9298b02c7930958ed52e0ccf4e4, which was discussed to be
ineffective, but left in place for the time being.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>