daemon: don't discard "invalid parameter" errors for archive endpoints
Commit ebcb7d6b406fe50ea9a237c73004d75884184c33 removed code that string-
matched errors to return the appropriate error-type. As part of that, it
defaulted to using a "System" (status 500) error.

We need to verify codepaths used by the archive functions, but let's
start with detecting invalid parameter errors, and preserve those.

This patch updates the code backing the following endpoints to return
a 400 (Invalid Parameter) error when present;

- `HEAD /containers/{name:.*}/archive`
- `GET /containers/{name:.*}/archive`
- `PUT /containers/{name:.*}/archive`

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>