daemon/links: remove newPortNoError utility
This utility was added in 12b6083c8f82db7e5db4c683cfe20151731ea851 as a
replacement for nat.NewPort(), which before that patch would panic on
invalid values, but was changed to return an error.

Given that the utility ignores any error, and these values are fixed values
for the test, let's remove it to simplify constructing the tests.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>