Move gateway-endpoint config to the Sandbox
Endpoint.updateExternalConnectivity is responsible for configuring
the network driver's external connectivity when a new gateway
endpoint is selected (including port mappings for the bridge
driver).

As part of that, it removes the gateway-ness of existing gateway
endpoints if necessary. But, that's the role of the Sandbox, one
Endpoint shouldn't be reconfiguring another.

Signed-off-by: Rob Murray <rob.murray@docker.com>